Fixes #21955 - improve repo sets api
 * searching
 * ability to query only enabled sets
 * adding arch and release_version to enabled repos list
 * support paging
