+++ app/lib/actions/katello/repository_set/enable_repository.rb
@@ -21,14 +21,14 @@ module Actions
 
         def repository_mapper(product, content, substituions, registry_name)
           if content.content_type == ::Katello::Repository::CANDLEPIN_DOCKER_TYPE
-            ::Katello::Candlepin::Content::DockerRepositoryMapper.new(product,
-                                                                content,
-                                                                registry_name)
+            ::Katello::Candlepin::DockerRepositoryMapper.new(product,
+                                                             content,
+                                                             registry_name)
 
           else
-            ::Katello::Candlepin::Content::RepositoryMapper.new(product,
-                                                                content,
-                                                                substituions)
+            ::Katello::Candlepin::RepositoryMapper.new(product,
+                                                       content,
+                                                       substituions)
           end
         end
       end
