+++ app/views/katello/api/v2/products/_product_content.json.rabl
@@ -1,14 +1,9 @@
 attribute :enabled, :product_id
-node :content do |pc|
-  c = pc.content
-  {
-    id: c.cp_content_id,
-    name: c.name,
-    label: c.label,
-    vendor: c.vendor,
-    type: c.content_type,
-    gpgUrl: c.gpg_url,
-    contentUrl: c.content_url,
-    modifiedProductIds: c.modified_product_ids
-  }
+
+glue :content do |_content|
+  attributes :name, :label, :vendor,
+             :cp_content_id => :id,
+             :content_type => :contentType,
+             :gpg_url => :gpgUrl,
+             :content_url => :contentUrl
 end
